Add New User window complete
Add New User window complete. Connection to SQLite database is now
working. But still shows "Unable to connect to database" or "Unable to
write to readonly database" error. This can easily be fixed if the
folder where the database is stored at is not set to "read-only" and
security folder has "Write" access.